Safari-611.1.10/LayoutTests/editing/async-clipboard/clipboard-write-text-requires-user-gesture-expected.txt

This test verifies that navigator.clipboard.writeText requires a user gesture.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Did not write to the clipboard.
PASS clipboardException.name is "NotAllowedError"
PASS successfullyParsed is true

TEST COMPLETE
